+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Fix :
  - remove test posts
+ - remove wip resume sections
 
 ### 0.0.1